% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_palette_colors.R
\name{get_palette_colors}
\alias{get_palette_colors}
\title{Get colors from specified color palette}
\usage{
get_palette_colors(color_palette = NULL, n_colors)
}
\arguments{
\item{color_palette}{chr name of color palette; if NULL selects random palette from `grDevices::hcl.pals()`}

\item{n_colors}{num how many colors to return}
}
\value{
chr vector of colors from palette
}
\description{
Get colors from specified color palette
}
