% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_random_clifford_attractor.R
\name{make_random_clifford_attractor}
\alias{make_random_clifford_attractor}
\title{Make a crazy Clifford attractor plot}
\usage{
make_random_clifford_attractor(point_color = "black", n_points = 1e+07,
  point_shape = 46, point_alpha = 0.01)
}
\arguments{
\item{point_color}{chr line color}

\item{n_points}{num length of vector to draw}

\item{point_shape}{num shape of point}

\item{point_alpha}{num alpha of point}
}
\value{
`ggplot` object
}
\description{
Make a crazy Clifford attractor plot
}
