% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_random_harmonograph.R
\name{make_random_harmonograph}
\alias{make_random_harmonograph}
\title{Draw a harmonograph using random data}
\usage{
make_random_harmonograph(n = 4, sequence = seq(0, 100, by = 0.001),
  point_size = 0.5, point_alpha = 0.5, ...)
}
\arguments{
\item{n}{num how many parameter terms}

\item{sequence}{num vector how many points to draw}

\item{point_size}{num size of point in `geom_point`}

\item{point_alpha}{num alpha of point in `geom_point`}

\item{...}{other args passed on to `geom_point`}
}
\value{
`ggplot` object
}
\description{
Draw a harmonograph using random data
}
\details{
inspired by and based on https://fronkonstin.com/2014/10/13/beautiful-curves-the-harmonograph/
}
