% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_random_scatterplot.R
\name{make_random_scatterplot}
\alias{make_random_scatterplot}
\title{Make a scatterplot of random points}
\usage{
make_random_scatterplot(points = 5000, point_shape = "square",
  background_color = NA, ...)
}
\arguments{
\item{points}{num number of points}

\item{point_shape}{chr shape of points}

\item{background_color}{chr background color;
defaults to `NA` which defers to default theme options}

\item{...}{other args passed to `geom_point()`, including
`alpha`, `size`, `color`, etc.}
}
\value{
`ggplot` object
}
\description{
Make a scatterplot of random points
}
\details{
Creates a data.frame of points length,
with column `x` drawn from random normal distribution,
column `y` from random uniform distribution,
and column `z` from random normal distribution.
}
