% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_random_stacked_area.R
\name{make_random_stacked_area}
\alias{make_random_stacked_area}
\title{Make random stacked area plot}
\usage{
make_random_stacked_area(groups = 20, n_colors = groups,
  color_palette = NULL,
  palette_colors = get_palette_colors(color_palette, n_colors))
}
\arguments{
\item{groups}{num how many groups to create}

\item{n_colors}{num how many different colors}

\item{color_palette}{chr color palette}

\item{palette_colors}{chr list of colors, usually generated from `get_palette_colors()`}
}
\value{
`ggplot` object
}
\description{
Make random stacked area plot
}
\details{
Idea and code from 'https://www.r-graph-gallery.com/137-spring-shapes-data-art.html'
}
