% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_random_stacked_density.R
\name{make_random_stacked_density}
\alias{make_random_stacked_density}
\title{Make a random stacked density plot}
\usage{
make_random_stacked_density(n_lines = 9, n_rows = 1000,
  n_colors = n_lines, color_palette = NULL,
  palette_colors = get_palette_colors(color_palette, n_colors), ...)
}
\arguments{
\item{n_lines}{num how many density lines}

\item{n_rows}{num how many points (lower will give more variability)}

\item{n_colors}{num how many colors to use; defaults to `n_lines`}

\item{color_palette}{chr which color palette to use}

\item{palette_colors}{chr vector of colors from color palette, usually generated by `get_palette_colors()`}

\item{...}{other arguments passed on to `layer()`}
}
\value{
`ggplot` object
}
\description{
Make a random stacked density plot
}
