% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_random_voronoi.R
\name{make_random_voronoi}
\alias{make_random_voronoi}
\title{Make a voronoi plot of random points}
\usage{
make_random_voronoi(shapes = sample(c(10, 100, 1000), 1),
  n_colors = round(shapes/2), gradient = TRUE, color_palette = NULL,
  palette_colors = get_palette_colors(color_palette, n_colors),
  fill_color = NULL, line_color = NULL, line_size = NULL)
}
\arguments{
\item{shapes}{num number of shapes to make}

\item{n_colors}{num number of colors to use in plot}

\item{gradient}{logical should color gradient be used}

\item{color_palette}{chr name of color palette to use}

\item{palette_colors}{chr list of colors, usually generated from `get_palette_colors()`}

\item{fill_color}{chr name of color to be used for fill}

\item{line_color}{chr name of color to be used for line}

\item{line_size}{num size of line width}
}
\value{
`ggplot` object
}
\description{
Make a voronoi plot of random points
}
\details{
Creates a data.frame of points length,
with column `x` drawn from random normal distribution,
column `y` from random uniform distribution,
and column `z` from random normal distribution.
}
