% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_arrrt.R
\name{theme_arrrt}
\alias{theme_arrrt}
\title{`ggplot` theme for `arrrt` objects}
\usage{
theme_arrrt(...)
}
\arguments{
\item{...}{args passed to `theme_void()`}
}
\value{
ggplot theme
}
\description{
`ggplot` theme for `arrrt` objects
}
